Move class to rearrange the packaging in tests
